+++ plugins/Pluginer.php
@@ -42,6 +42,11 @@ class Pluginer extends Admin
 
 	// applyPlugin
 
+	public function init(): void
+	{
+		$this->applyPlugin(__FUNCTION__, func_get_args());
+	}
+
 	public function name()
 	{
 		return $this->applyPlugin(__FUNCTION__, func_get_args());